src/presentation/pages/survey-list/survey-list.spec.tsx
import { render, screen, waitFor } from '@testing-library/react'
import { SurveyList } from './survey-list'
import { ILoadSurveyList } from '@/domain/usecases'
import { SurveyModel } from '@/domain/models'
import { mockLoadSurveyList } from '@/domain/test'

class LoadSurveyListSpy implements ILoadSurveyList {
  callsCount: number = 0
  surveys = mockLoadSurveyList()

  async loadAll(): Promise<SurveyModel[]> {
    this.callsCount++
    return this.surveys
  }
}

type SutTypes = {
  loadSurveyListSpy: LoadSurveyListSpy
}

const makeSut = (): SutTypes => {
  const loadSurveyListSpy = new LoadSurveyListSpy()
  render(<SurveyList loadSurveyList={loadSurveyListSpy} />)
  return { loadSurveyListSpy }
}

describe('Name of the group', () => {
  test('Should present 4 empty items on start', async () => {
    makeSut()
    const surveyList = screen.getByTestId('survey-list')
    expect(surveyList.querySelectorAll('li:empty')).toHaveLength(4)
    await waitFor(() => surveyList)
  })

  test('Should call LoadSurveyList', async () => {
    const { loadSurveyListSpy } = makeSut()
    expect(loadSurveyListSpy.callsCount).toBe(1)
    await waitFor(() => screen.getByRole('heading'))
  })

  test('Should render CardSurvey on success', async () => {
    makeSut()
    const surveyList = screen.getByTestId('survey-list')
    await waitFor(() => surveyList)
    expect(surveyList.querySelectorAll('li.survey_wrap')).toHaveLength(3)
  })
})
